+++ org/antlr/v4/runtime/CommonToken.cpp
@@ -103,7 +103,7 @@ namespace org {
                         return L"";
                     }
                     size_t n = input->size();
-                    if (start < n && stop < n) {
+                    if ((size_t)start < n && (size_t)stop < n) {
                         return input->getText(misc::Interval::of(start,stop));
                     } else {
                         return L"<EOF>";
